+++ crates/threshold-signature-server/src/user/api.rs
@@ -34,7 +34,9 @@ use entropy_kvdb::kv_manager::{
 };
 use entropy_protocol::SigningSessionInfo;
 use entropy_protocol::ValidatorInfo;
-use entropy_shared::{types::KeyVisibility, OcwMessageDkg, X25519PublicKey, SIGNING_PARTY_SIZE};
+use entropy_shared::{
+    types::KeyVisibility, HashingAlgorithm, OcwMessageDkg, X25519PublicKey, SIGNING_PARTY_SIZE,
+};
 use futures::{
     channel::mpsc,
     future::{join_all, FutureExt},
@@ -67,7 +69,7 @@ use crate::{
         substrate::{
             get_program, get_registered_details, get_subgroup, return_all_addresses_of_subgroup,
         },
-        user::{check_in_registration_group, do_dkg, send_key},
+        user::{check_in_registration_group, compute_hash, do_dkg, send_key},
         validator::get_signer,
     },
     signing_client::{ListenerState, ProtocolErr},
@@ -88,6 +90,8 @@ pub struct UserSignatureRequest {
     pub validators_info: Vec<ValidatorInfo>,
     /// When the message was created and signed
     pub timestamp: SystemTime,
+    /// Hashing algorithm to be used for signing
+    pub hash: HashingAlgorithm,
 }
 
 /// Type for validators to send user key's back and forth
@@ -142,14 +146,36 @@ pub async fn sign_tx(
 
     let message = hex::decode(&user_sig_req.message)?;
     let auxilary_data = user_sig_req.auxilary_data.as_ref().map(hex::decode).transpose()?;
-    let message_hash = Hasher::keccak(&message);
-    let message_hash_hex = hex::encode(message_hash);
-    let subgroup_signers = get_current_subgroup_signers(&api, &rpc, &message_hash_hex).await?;
+
+    let program = get_program(&api, &rpc, &user_details.program_pointer).await?;
+    let mut runtime = Runtime::new();
+    let signature_request = SignatureRequest { message, auxilary_data };
+    runtime.evaluate(&program, &signature_request)?;
+
+    // We decided to do Keccak for subgroup selection for frontend compatability
+    let message_hash_keccak = compute_hash(
+        &HashingAlgorithm::Keccak,
+        &mut runtime,
+        program.as_slice(),
+        signature_request.message.as_slice(),
+    )?;
+    let message_hash_keccak_hex = hex::encode(message_hash_keccak);
+
+    let subgroup_signers =
+        get_current_subgroup_signers(&api, &rpc, &message_hash_keccak_hex).await?;
     check_signing_group(&subgroup_signers, &user_sig_req.validators_info, signer.account_id())?;
 
     // Use the validator info from chain as we can be sure it is in the correct order and the
     // details are correct
     user_sig_req.validators_info = subgroup_signers;
+
+    let message_hash = compute_hash(
+        &user_sig_req.hash,
+        &mut runtime,
+        program.as_slice(),
+        signature_request.message.as_slice(),
+    )?;
+    let message_hash_hex = hex::encode(message_hash);
 
     let signing_session_id = SigningSessionInfo {
         account_id: SubxtAccountId32(*signed_msg.account_id().as_ref()),
@@ -160,13 +186,6 @@ pub async fn sign_tx(
     if !has_key {
         recover_key(&api, &rpc, &app_state.kv_store, &signer, signing_address).await?
     }
-
-    let program = get_program(&api, &rpc, &user_details.program_pointer).await?;
-
-    let mut runtime = Runtime::new();
-    let signature_request = SignatureRequest { message, auxilary_data };
-
-    runtime.evaluate(&program, &signature_request)?;
 
     let (mut response_tx, response_rx) = mpsc::channel(1);